gevent_socketio_flask/requirements.txt
Flask==0.9
Jinja2==2.6
Pygments==1.5
Werkzeug==0.8.3
distribute==0.6.27
docutils==0.9.1
gevent==0.13.7
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.0
wsgiref==0.1.2
